Speicalize `copy!` for triangular, and use `copy!` in `ldiv` (#1263)
Currently, we specialize `copyto!` for a triangular source. However,
this has two branches, depending on whether the axes match. In the
branch where the axes do match, we may use `copy!` instead, and
specialize this in terms of the internal `_copyto!` which is identical
in implementation.

We also use `copy!` in `ldiv` instead of `copyto!`. These should be
equivalent in most cases, barring one extra axes check in `copy!`, as
the fallback method for `copy!` calls `copyto!` internally. However, the
advantage comes for triangular matrices, where `copy!` doesn't have
branches, as the axes necessarily match.

As a consequence, this reduces the TTFX in operations like
```julia
julia> using Random, LinearAlgebra

julia> A = rand(4,4);

julia> @time A \ UpperTriangular(A);
  0.598575 seconds (1.28 M allocations: 61.788 MiB, 98.57% compilation time: 3% of which was recompilation) # master
  0.487267 seconds (1.01 M allocations: 49.726 MiB, 3.54% gc time, 98.95% compilation time) # this PR
```
```julia
julia> @time UpperTriangular(A) / A;
  0.826212 seconds (1.45 M allocations: 71.427 MiB, 16.96% gc time, 84.47% compilation time) # master
  0.616258 seconds (1.28 M allocations: 63.135 MiB, 2.65% gc time, 99.19% compilation time) # this PR
```

I've renamed the internal functions as `_copyto!` -> `_copy!` and
`copyto_unaliased!` -> `copy_unaliased!`, as these are closer to the
meaning of `copy!` than to `copyto!`.