Clean up `herk_wrapper!` and add 5-arg tests (#1254)
This method can only be reached with complex eltypes, and after
promotion of alpha and beta, they cannot be `Bool` anymore. Also, this
adds 5-arg `mul!` of which I'm not sure we had some (even though
coverage said it was covered, which is strange because it shouldn't due
to the issue fixed in #1247).