Unaliasing and short-circuiting in `copytrito!` (#1287)
The LAPACK function `lacpy!` seems to handle aliasing correctly already,
but since the LAPACK documentation doesn't expect the output arguments
to be aliased, it's best if we carry out the unaliasing by ourselves.

I've also added a quick return criterion if the two matrices are
identical, in which case there is nothing to copy. This would be useful
in copying elements between a triangular/symmetric matrix and its
parent.