Inline `generic_matmatmul!` branch in strided triangular matmul (#1262)
For combinations of strided matrices and strided triangular matrices, we
would end up taking the methods defined in `LinearAlgebra`, so we may
avoid the constant-propagation and hardcode the `_generic_matmatmul!`
call. This improves TTFX, as the no-op but expensive-to-compile `wrap`
call is elided.

```julia
julia> using LinearAlgebra

julia> A = zeros(4,4);

julia> @time A * UpperTriangular(A);
  0.458913 seconds (1.22 M allocations: 59.769 MiB, 51.63% gc time, 97.84% compilation time: 4% of which was recompilation) # master
  0.077198 seconds (174.52 k allocations: 8.683 MiB, 92.75% compilation time) # this PR
```