Band indexing for adj/trans (#1299)
The advantage of this is that we may forward the indexing to the parent,
and if the parent is a structured matrix and the band is a compile-time
constant, the value may be evaluated as a constant.

E.g.:
```julia
julia> A = UnitUpperTriangular(reshape(1:9, 3, 3))
3×3 UnitUpperTriangular{Int64, Base.ReshapedArray{Int64, 2, UnitRange{Int64}, Tuple{}}}:
 1  4  7
 ⋅  1  8
 ⋅  ⋅  1

julia> function f(A, i, ::Val{band}) where {band}
           x = Adjoint(A)[LinearAlgebra.BandIndex(band,i)]
           Val(x)
       end
f (generic function with 1 method)

julia> @inferred f(A, 1, Val(0))
Val{1}()

julia> @inferred f(A, 1, Val(1))
Val{0}()
```
In this case, the band index `0` is forwarded to the parent
(`UnitUpperTriangular`), and the `getindex` is evaluated at compile-time
to return `1` (if within bounds). Similarly, the band index `1`
corresponds to the zeros of the triangular matrix, so the value is
evaluated at compile-time. This may help with removing branches in
indexing.