Only `@noinline` error path in `matmul_size_check` (#1310)
Instead of `@noinline` on the entire size-check function, we now
separate the error-throwing part into a separate function and mark it as
`@noinline`. This way, the size check may still be evaluated inline, and
only the error path will not be inlined.

This improves performance for small matmul.

```julia
julia> A = [1 2; 3 4];

julia> @btime $A * $A;
  53.361 ns (2 allocations: 112 bytes) # v"1.13.0-DEV.438"
  47.504 ns (2 allocations: 112 bytes) # this PR
```