add missing methods for division of Hessenberg matrices (#1322)
As noted [on
discourse](https://discourse.julialang.org/t/add-a-upperhessenberg-row-in-linearalgebra-factorize/128530/4?u=stevengj),
we were missing `/` and `\` methods for `UpperHessenberg` matrices,
despite the existence of optimized `ldiv!` and `rdiv!` methods, so it
was falling back to slow $O(n^3)$ methods.

While I was at it, I also added methods for transpose/adjoint Hessenberg
matrices.

---------

Co-authored-by: Jishnu Bhattacharya <jishnub.github@gmail.com>