`Char` uplo in `Bidiagonal` constructor (#1342)
This is an internal constructor, but many methods involving `Bidiagonal`
matrices currently use the `Char` `uplo` directly to construct a new
`Bidiagonal` matrix. We therefore may require the constructors to accept
a `Char` as well, wherever they accept a `Symbol`. This already works
for some of the `Bidiagonal` constructors, and this PR plugs a missing
case.

Fixes issues like
```julia
julia> B = Bidiagonal(fill(SMatrix{2,3}(1:6), 4), fill(SMatrix{1,3}(1:3), 3), :L)
4×4 Bidiagonal{StaticArraysCore.SArray{S, Int64, 2} where S<:Tuple, Vector{StaticArraysCore.SArray{S, Int64, 2} where S<:Tuple}}:
 [1 3 5; 2 4 6]        ⋅               ⋅               ⋅       
 [1 2 3]         [1 3 5; 2 4 6]        ⋅               ⋅       
       ⋅         [1 2 3]         [1 3 5; 2 4 6]        ⋅       
       ⋅               ⋅         [1 2 3]         [1 3 5; 2 4 6]

julia> 2B
ERROR: MethodError: no method matching Bidiagonal(::Vector{SMatrix{2, 3, Int64, 6}}, ::Vector{SMatrix{1, 3, Int64, 3}}, ::Char)
The type `Bidiagonal` exists, but no method is defined for this combination of argument types when trying to construct it.

Closest candidates are:
  Bidiagonal(::Vector{T}, ::Vector{S}, ::Symbol) where {T, S}
   @ LinearAlgebra ~/Dropbox/JuliaPackages/LinearAlgebra_2.jl/src/bidiag.jl:77
  Bidiagonal(::V, ::V, ::AbstractChar) where {T, V<:AbstractVector{T}}
   @ LinearAlgebra ~/Dropbox/JuliaPackages/LinearAlgebra_2.jl/src/bidiag.jl:71
  Bidiagonal(::V, ::V, ::Symbol) where {T, V<:AbstractVector{T}}
   @ LinearAlgebra ~/Dropbox/JuliaPackages/LinearAlgebra_2.jl/src/bidiag.jl:68
  ...
```
After this,
```julia
julia> 2B
4×4 Bidiagonal{StaticArraysCore.SArray{S, Int64, 2} where S<:Tuple, Vector{StaticArraysCore.SArray{S, Int64, 2} where S<:Tuple}}:
 [2 6 10; 4 8 12]        ⋅                 ⋅                 ⋅       
 [2 4 6]           [2 6 10; 4 8 12]        ⋅                 ⋅       
       ⋅           [2 4 6]           [2 6 10; 4 8 12]        ⋅       
       ⋅                 ⋅           [2 4 6]           [2 6 10; 4 8 12]
```

We may require that each method passes a `Symbol` as `uplo`, but this
would be a much larger change, and the current approach seems to be less
disruptive.