slight optimization of exp, sqrt, ^ for positive semidefinite output (#1359)
The speedup is small, my benchmarks show 2-5%, but since these functions
are so often used I think it's worth it. Conversely, the same
optimization applies to `cosh`, `acosh`, `acos`, and the positive part
of other functions but I didn't do it because I don't think the speedup
is worth the complexity.

---------

Co-authored-by: Steven G. Johnson <stevenj@alum.mit.edu>