Forward `_tri(l/r)div!` to parent of triangular (#1348)
These loop over the stored triangular halves, so we may forward the
operations to the parents. These are also internal methods that are not
applied to unit triangular matrices, so we restrict these to
`UpperTriangular` and `LowerTriangular` matrices in this PR. This makes
it straightforward to forward the operations to the parent.