Change default symmetriceigen algorithm back to RobustRepresentations (#1363)
See #1313 

A brief summary: RobustRepresentations (`LAPACK.syevr!`) was the only
algorithm implemented until Julia v1.11. Then,
https://github.com/JuliaLang/julia/pull/49355/ implemented interface to
other algorithms, and changed the default to DivideAndConquer
(`LAPACK.syevd!`) based on its better numerical accuracy and
performance. But, it turned out that in some LAPACK implementation, the
DivideAndConquer fails more frequently than RobustRepresentations
(#1313). Based on the discussion in #1313 , this PR reverts the default
algorithm to RobustRepresentations.

Once the problem with the RobustRepresentations in the problematic
LAPACK implementation has been sorted out, the default could be changed
to DivideAndConquer.