Make `adjoint`/`transpose` for `Diagonal`/`Bidiagonal` lazy (#1368)
Currently, `adjoint` and `transpose` for a block `Diagonal` and a block
`Bidiagonal` matrix allocates the diagonals. This was not intentional,
and we may use lazy wrappers instead that ensures that the result shares
memory with the original array, which is what the documentation states.
After this PR, mutating the `adjoint` will automatically mutate the
parent array.