Check for elementwise equality when comparing with `UniformScaling` (#1369)
This works around issues with block matrices like:
```julia
julia> A = Diagonal(fill(I(2), 4))
4×4 Diagonal{Diagonal{Bool, Vector{Bool}}, Vector{Diagonal{Bool, Vector{Bool}}}}:
 Diagonal(Bool[1, 1])  …        ⋅       
       ⋅                        ⋅       
       ⋅                        ⋅       
       ⋅                  Diagonal(Bool[1, 1])

julia> A == I
true

julia> A[1,1] == I[1,1]
false
```
The checks in this PR ensure that if a matrix is `==` a
`UniformScaling`, their elements are equal as well.

This builds on top of
https://github.com/JuliaLang/LinearAlgebra.jl/pull/1367, and may be
rebased once that is merged.