Generic path in `ldiv!` for `Diagonal` (#1261)
The performance improvement in `ldiv!` is best seen using
`SparseArrays`, as this seems to lack a specialized `ldiv!` method
currently:
```julia
julia> D = Diagonal(rand(3000));

julia> S = sprand(size(D,1), 0.01);

julia> @btime ldiv!($D, $S);
  29.620 μs (0 allocations: 0 bytes) # master
  15.388 μs (22 allocations: 154.47 KiB) # This PR
```
This is because the broadcasting method is taken. Unfortunately, in this
example, the in-place broadcasting for a sparse array is allocating, but
this should be allocation-free in general.