Add 'eigmin'/'eigmax' methods for 'Eigen' (#1468)
We add the methods 'eigmin' and 'eigmax' on the type 'F::Union{Eigen,
GeneralizedEigen}' similarly to how 'eigvals' is defined on 'F'. Note
that the corresponding unit tests in 'test/eigen.jl' that call 'eigvals'
on such a struct are updated to include 'eigmin' and 'eigmax' calls as
well; if all eigenvalues are real, the actual functionality of the
methods is tested, and if any happen to be complex, then we test that
they throw 'MethodError's.

(Resolves #1464 by @longemen3000.)

---------

Co-authored-by: Jishnu Bhattacharya <jishnub.github@gmail.com>