Add missing methods to `diagind` documentation (#1473)
Previously, the `diagind` documentation only covered the methods that
take the full matrix as input, neglecting the methods that accept
dimensions in lieu despite both styles being publicly exported. This
update edits the docstring accordingly to reflect the missing methods,
also adding an additional example block.

Additionally, this moves the dimension-based methods from above the
matrix-based methods to below them, since the docstring is attached to
the matrix-based method definition. For consistency, it further
rearranges the order of the three dimension-based methods, moving the
one-liners below the multi-liner just as is already the case for the
matrix-based method definitions.

(Resolves #1469.)